slightly refactor reader fixture
